* fixed sorting culler issue with nested effects * fixed state id assignment bug * fixed state handling of local effects * clean up

+++ trunk/Wire/Renderer/GXRenderer/WireGXIndexBuffer.cpp
@@ -26,8 +26,8 @@ PdrIndexBuffer::PdrIndexBuffer(Renderer* pRenderer, const IndexBuffer*
 //----------------------------------------------------------------------------
 PdrIndexBuffer::~PdrIndexBuffer()
 {
-	TArray<TMap<UShort, PdrDisplayList*>::MapElement>* pArray =
-		mDisplayLists.GetArray();
+	TArray<TMap<UInt, PdrDisplayList*>::MapElement>* pArray = mDisplayLists.
+		GetArray();
 
 	for (UInt i = 0; i < pArray->GetQuantity(); i++)
 	{
